+++ seantis/dir/events/tests/test_import.py
@@ -206,38 +206,38 @@ class TestImport(IntegrationTestCase):
             ids = ['event1', 'event2', 'event3', 'event4',
                    'event5', 'event6', 'event7', 'event8']
             events = from_ids(ids[:4])
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 4)
             imported = [i.getObject().source_id for i in self.catalog.query()]
             self.assertEquals(ids[:4], imported)
 
             # Import with limit
             events = from_ids(ids[4:])
-            imports = importer.fetch_one('source', fetch, limit=2)
+            imports, deleted = importer.fetch_one('source', fetch, limit=2)
             self.assertEquals(imports, 2)
             self.assertEquals(len(self.catalog.query()), 6)
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 2)
             self.assertEquals(len(self.catalog.query()), 8)
 
             # Force reimport
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 0)
             self.assertEquals(len(self.catalog.query()), 8)
-            imports = importer.fetch_one('source', fetch,
+            imports, deleted = importer.fetch_one('source', fetch,
                                                   reimport=True)
             self.assertEquals(imports, 4)
             self.assertEquals(len(self.catalog.query()), 8)
 
             # Reimport updated events
             events = from_ids(ids)
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 8)
             self.assertEquals(len(self.catalog.query()), 8)
 
             # Test import of given source IDs only
             events = from_ids(ids)
-            imports = importer.fetch_one(
+            imports, deleted = importer.fetch_one(
                 'source', fetch, source_ids=ids[2:6]
             )
             self.assertEquals(imports, 4)
@@ -267,7 +267,7 @@ class TestImport(IntegrationTestCase):
                 cat1=set(), cat2=set(['cat2-1', 'cat2-2', 'cat2-3'])
             ))
 
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 3)
             self.assertTrue('cat1-1' in self.directory.cat1_suggestions)
             self.assertTrue('cat1-2' in self.directory.cat1_suggestions)
@@ -315,7 +315,7 @@ class TestImport(IntegrationTestCase):
             # event['attachment_1'] =
             # event['attachment_2'] =
 
-            imports = importer.fetch_one('source', lambda: [event])
+            imports, deleted = importer.fetch_one('source', lambda: [event])
             imported = self.catalog.query()
             self.assertEquals(imports, 1)
             self.assertEquals(len(imported), 1)
@@ -357,8 +357,9 @@ class TestImport(IntegrationTestCase):
                                                   fetch_id='fetch_id'))
             events.append(self.create_fetch_entry(source_id='source_id_2',
                                                   fetch_id='fetch_id'))
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 3)
+            self.assertEquals(deleted, 0)
             self.assertEquals(len(self.directory.keys()), 3)
 
             # Second import
@@ -367,9 +368,10 @@ class TestImport(IntegrationTestCase):
             events.append(self.create_fetch_entry(source_id='source_id_3',
                                                   fetch_id='fetch_id'))
 
-            imports = importer.fetch_one('source', fetch,
+            imports, deleted = importer.fetch_one('source', fetch,
                                                   autoremove=True)
             self.assertEquals(imports, 1)
+            self.assertEquals(deleted, 2)
 
             ids = [item.source_id for id, item in self.directory.objectItems()]
             self.assertEquals(len(ids), 2)
@@ -384,7 +386,7 @@ class TestImport(IntegrationTestCase):
             # Import event
             importer = ExternalEventImporter(self.directory)
             event = self.create_fetch_entry(source_id='s', fetch_id='f')
-            imports = importer.fetch_one('source', lambda: [event])
+            imports, deleted = importer.fetch_one('source', lambda: [event])
             self.assertEquals(imports, 1)
 
             # Hide event
@@ -396,9 +398,9 @@ class TestImport(IntegrationTestCase):
             hidden.hide()
 
             # Re-import event
-            imports = importer.fetch_one('source', lambda: [event])
+            imports, deleted = importer.fetch_one('source', lambda: [event])
             self.assertEquals(imports, 0)
-            imports = importer.fetch_one(
+            imports, deleted = importer.fetch_one(
                 'source', lambda: [event], reimport=True
             )
             self.assertEquals(imports, 1)
@@ -422,7 +424,7 @@ class TestImport(IntegrationTestCase):
             # Import event
             importer = ExternalEventImporter(self.directory)
             event = self.create_fetch_entry(source_id='s', fetch_id='f')
-            imports = importer.fetch_one('source', lambda: [event])
+            imports, deleted = importer.fetch_one('source', lambda: [event])
             self.assertEquals(imports, 1)
 
             # Add own event